Handle default flag value properly
Cabal doesn't expose default flag values the normal way (apparently
this is a longstanding bug/omission). The user probably shouldn't be
enabling LiquidHaskell by default anyway, but just in case - to
minimize surprising behavior - we load in the package description and
extract the default flag value if it isn't specified via
configuration.